Move readable duration method into `@joincivil/utils`
