+++ packages/utils/src/index.ts
@@ -2,3 +2,4 @@ import "./rxjs";
 export * from "./language";
 export * from "./civilHelpers";
 export * from "./contracts";
+export * from "./viewHelpers";
